Implement consensus taxonomy analysis
